[BigQuery] Remove last reference to the state dataset in views (Recidiviz/recidiviz-data#33467)
## Description of the change

The final BQ view that was referencing the `state` dataset was a
validation that checks that each table has unique primary keys ACROSS
states. I updated this view to just enforce this invariant for
`normalized_state` tables.

I don't think it's a big deal to drop this validation for the `state`
dataset because we already have validation in ingest pipelines that
within a state the primary keys are unique, and also nobody queries from
the `state` anymore.

This change will make it easier to move the `update_state_dataset` step
into the view graph (will simplify how I deal with unused view
exemptions).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#29531
Part of Recidiviz/recidiviz-data#29515

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 655963742d7f2a6c277b63213fb5bf098856d1cd