[US_AZ] Ingest `state_sentence_status_snapshot` (Recidiviz/recidiviz-data#33000)
## Description of the change

### What this does
This PR adds a view to hydrate `state_sentence_status_snapshot`. It also
updates sentence external IDs across all views where sentences are
generated so that they have the form `OFFENSE_ID-COMMITMENT_ID-DOC_ID`,
since sentence_group external IDs have the form `COMMITMENT_ID-DOC_ID`.
This will make the two entities easier to relate visually.

AZ has statuses associated with specific sentences in the raw data, but
they do not perfectly map to the statuses in our schema. The biggest
difference is that the only "completed" statuses are those that are
completed early, vacated, or commuted. There are no standard `COMPLETED`
statuses - all sentences that have presumably been completed on the
expected schedule are simply marked as `IMPOSED`. This view uses the
`SENTENCE_END_DTM` field associated with the sentence - which is meant
to hold the final end date of the sentence - to determine if a sentence
with the `IMPOSED` status should be mapped to `SERVING` or `COMPLETED`.

This view treats historical sentences differently than current
sentences. This is because when Arizona completed a system migration in
late 2019, the CREATE_DTM and UPDT_DTM values on many tables were
overwritten to reflect the date of the migration rather than the date
the row was created in the old system.

As a proxy for update datetimes for historical sentence statuses, we use
sentence start and end datetimes that are documented separately. For
sentences that began after the system migration, we can rely on the
CREATE_DTM and UPDT_DTM fields.

ADCRR staff, or the system itself, also seems to update offense
information indefinitely for years after the associated sentence has
ended. The sentence status is not what is getting updated in these
changes, so this view discounts any "updates" that are made after the
end of a sentence.

### Tests:
- [x] run mappings script parses successfully
- [x] sandbox ingest + normalization
[pipeline](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-09-12_12_17_05-2336565431135822698?authuser=0&organizationId=448885369991&project=recidiviz-staging)
succeeds

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#32613

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 24018eafa5044ba9037fd8db64d12b71de450e26