[US_PA] Quick fix for new state staff roster format (Recidiviz/recidiviz-data#33518)
## Description of the change

This PR puts in a quick fix for the new state staff roster format (which
includes slightly different names for location descriptions). This PR
revises it so that we use location code directly from the roster in
cases where the location name doesn't merge on with the existing
reference file. In addition, this PR adds in a new enum value for
caseload type.

Longer term refactor of state staff to account for the new roster format
(which may include staff without employee ids) here:
https://github.com/Recidiviz/recidiviz-data/pull/33515

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9aace88f0fb9e8da600ddffc1c15454eebf1a8f0