remove MI district transformation from current_staff query (Recidiviz/recidiviz-data#33528)
## Description of the change

Remove transformation of MI districts so that we get the full district
10 information. cc @mayukas

This will affect both the workflows staff record and admin panel, so
it's possible that we have a small bit of time where people's districts
don't match exactly since the workflows ETL and roster sync ETL don't
run at the exact same time, but I think if we warn Kim to look out for
these and potentially put up an intercom banner we should be safe.

In addition to the above, there are 33 users whose district in
user_override is 10, 1a, 4a, or 4b, which means that when this change
runs, they would not be updated.
* Of these users, 7 of them have a district in the roster table- some of
them match so we can remove the override, and some of them don't match
but I checked them all and they all _used_ to be in the override
district, so I think they just moved and we never updated the admin
panel, so I'm going to remove the override for those too.
* Of the rest, only 3 have routes in the admin panel: 2 of these are in
district 10 and so we need to find out their actual district, and the
last one is in district 1a so we just need to capitalize it at around
the same time as roster sync runs. One of those users has never logged
in and can probably be deleted outright if we don't know their actual
district, and the others haven't logged in in 5 and 9 months
respectively so I'm not concerned with getting the timing super precise.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#25398

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 76af6b1a0bc2c82bd496a6f371ccb906490af8d0