Switch time periods to count backwards from end date instead of counting forwards from start date (Recidiviz/recidiviz-data#33315)
## Description of the change

The current behavior is that we generate time periods by starting at the
user's specified start date and counting forward until we have periods
that cover the user's entire specified range. For example, if the user
specifies the start date 5/1/24, the end date 6/15/24, and period of
WEEK, we generate the following time periods:

| start_date | end_date |
| ---------- | ---------- |
| 05/01/24  | 05/08/24 |
| ... | ... |
| 06/12/24 | 06/19/24 |

Notice that the last time period overhangs the user's specified end date
of 06/15. This happens because we specify that all periods will be the
same length (a necessary condition for valid comparisons of time
periods).

In order to better align with users' expectations, this PR changes the
behavior to construct time periods by counting **backwards from the
specified end date** _instead of_ **forwards from the specified start
date**.

For the above example, this new behavior results in the following time
periods:

| start_date | end_date |
| ---------- | ---------- |
| 04/27/24 | 05/04/24 |
| ... | ... |
| 06/08/24 | 06/15/24 |

One period still "overhangs" the user's specified date range, but now
this overhanging period is at the beginning of the date range instead of
the end.

We had considered eliminating overhanging periods to mitigate confusion
around this, but the problem with that is when we calculate rolling
metrics, we need to **look back further than the user's specified start
date**. (For example, if the user wants monthly rolling metrics
calculated daily for all days in their specified date range, many of
those data points need to be calculated over time periods that start
before the user's specified start date.)

If we still find it too confusing to have these overhanging periods
appear in dashboards, we can modify individual visualizations to not
show those periods, as Mayuka suggested.

See the accompanying Looker PR here:
https://github.com/Recidiviz/looker/pull/627

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://github.com/Recidiviz/recidiviz-data/issues/32406

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e0b1aafd8a08049e42bbddd4cd5149967c283c27