[US_ME] Change the supervision violation criteria (Recidiviz/recidiviz-data#33477)
## Description of the change

We changed the supervision violation criteria from only including folks
with 'FELONIES' or 'MISDEMEANORS' to any violation.

I uploaded it to sandbox (under `hsalas`) and it all looks as expected.
This reduced the number of people eligible from 1102 to 1003.

## Related issues

Closes Recidiviz/recidiviz-data#32956

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a64cff0bd2303cbc1b6be54cb0ce27e720145bcf