[raw data import] Don't parse filenames that have timestamps using underscores (Recidiviz/recidiviz-data#33533)
## Description of the change

Previously the timestamp regex matched timestamps using underscores, so
the `is_path_normalized` function would return true for a filename with
format `unprocessed_2024-09-13T17_19_55_995516_raw_ContactNoteType.csv`,
then when we tried to get `filename_parts_from_path` for that path we
would get error
```
utc_upload_datetime = datetime.datetime.fromisoformat(
^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
ValueError: Invalid isoformat string: '2024-09-13T17_19_55_995516'
```

if we're expecting a file to be named
`{PROCESSED_PREFIX}_{ISO_DATETIME}_raw_{FILE_TAG}{SUFFIX}.csv` then i
dont think we should consider a filename with
`2024-09-13T17_19_55_995516` as a valid normalized filename.

the `ingest_filename_normalization` cloud function checks if a path is
normalized by checking that the file has the prefix `unprocessed` or
`processed`. so a filename that has that prefix but the rest of the file
name format is incorrect is ignored by the cloud function and should be
ignored by the import dag

We could as a future AI have `ingest_filename_normalization` cf modify
'2024-09-13T17_19_55_995516' -> '2024-09-13T17:19:55:995516' if this is
a case we see often

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#33529

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e6552497fba45b39927e08341fa2bfe0822ce293