[Justice Counts] Don't write to file to standardize Excel Workbook  (Recidiviz/recidiviz-data#33509)
## Description of the change

This PR addresses an issue related to handling CSV uploads via Bulk
Upload. OCJS was not able to upload a CSV file earlier today and a [
"truncated file header" error
](https://recidiviz-inc.sentry.io/issues/5802274479/?project=4504532096516096&query=is%3Aunresolved+issue.priority%3A%5Bhigh%2C+medium%5D&referrer=issue-stream&stream_index=1)was
surfaced to Sentry.

The key adjustment made in this PR was removing the logic in
`WorkbookStandardizer.upload_workbook` where the excel workbook is
overwritten. This was causing errors because, when we tried to read from
the excel_file while it was open as a stream, it was leading to
interference with the stream’s state. This resulted in the file being
improperly accessed, causing the "truncated file header" error later on.
By removing this logic and returning the a dictionary that maps sheet
name to sheet dataframe (the same data structure needed by
WorkbookUploader). With this change I was able to upload the problematic
sheet.

## Related issues

Closes Recidiviz/recidiviz-data#33505

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3e03ccdd17a24042dc49226be84d7d7d9c80ca77