[US_AZ] Refactor state_sentence_status_snapshot (Pre-migration only) (Recidiviz/recidiviz-data#33547)
## Description of the change

This is the first part of a two-part refactor that will incorporate
pre-migration and post-migration sentence status snapshots into this
ingest view independently of one another for simplicity. I am
refactoring this view because the original approach violated constraints
in ingest - see ingest pipeline failure
[here](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-09-20_11_11_01-11666868511239182787;bottomTab=JOB_LOGS;expandBottomPanel=true;logsSeverity=ERROR;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))).

In the original approach, we were overwriting the
`status_update_datetime` of `Imposed` statuses that happened
pre-migration to be the `sentence_begin_dtm`, and we were creating
`Recidiviz Marked Completed` statuses using the `sentence_end_dtm` as a
`status_update_datetime`. The constraint is violated because sometimes
these two dates are identical, so there are multiple distinct statuses
with identical `status_update_datetime` values.

This refactor is concerned only with pre-migration sentences. It uses
the `sentence_begin_dtm` as a `status_update_datetime` for sentences
that started before 2019-11-30. It creates `Recidiviz Marked Completed`
statuses using the `sentence_end_dtm` as a `status_update_datetime`,
_unless_ the `sentence_end_dtm` is equal to the `sentence_begin_dtm`.
When that is the case, the `UPDT_DTM` value from the raw data is used
instead.

Tests:
- [x] run_mappings_against_ingest_view_results script parses
successfully

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d878d67456d30c868d36288ac4f42fa0877aad2f