[Cost] Optimize tes_utils.py functions to query from task-specific views (Recidiviz/recidiviz-data#33549)
## Description of the change

I was digging into some BQ cost graphs with @colincadams and on days
where we saw super spiky non-service account costs, the
`get_task_query()` query was often the culprit. This function is
presumably only run in Python notebooks and the query logic is not
referenced by any view, which is why our tests that enforce that queries
do not reference the `all_tasks_materialized` directly were not
triggered.

This PR optimizes that function to directly query the task-specific
table rather than query `all_tasks_materialized`, which does an
expensive scan over a bunch of data we don't need.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

N/A

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 06c9c54897f3ec35f625c92b0f152887d19c27a4