[Airflow] Move update_normalized_state_dataset logic into BQ view graph (Recidiviz/recidiviz-data#33480)
## Description of the change

Similar to Recidiviz/recidiviz-data#32621, moves the `update_normalized_state_dataset` process
into the view graph. We now generate the `normalized_state` dataset by
materializing a set of views in the `normalized_state_views` dataset.

As part of this change, I had to make a number of changes to the
calculation documentation generation code so that the view / product
docs ended up mostly the same. The docs generation code is already a
mess, so not my most elegant work...

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#29519

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5bd7758f50cf8e2b5739dae4aa6c8063ea145daa