[Cleanup] Rename us_xx_normalized_state_new datasets back to us_xx_normalized_state (Recidiviz/recidiviz-data#33550)
## Description of the change

Renames the `us_xx_normalized_state_new` datasets back to
`us_xx_normalized_state` now that the legacy `us_xx_normalized_state`
datasets have been deleted. This is the final cleanup step for combined
ingest and normalization pipelines!

After this PR is deployed to staging, I will run a script to drop all
the `us_xx_normalized_state_new` datasets and then will as a reminder to
do the same in prod in go/deploy-checklist.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#31741
Closes Recidiviz/recidiviz-data#29517

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: edf29b4ea3f2d4136fd368ecfa07dbcc39ee5b38