[US_AZ] Only store most recent home plan status (Recidiviz/recidiviz-data#33611)
## Description of the change

There were a lot of inconsistencies and inaccuracies in the home plan
preprocessed view that pulls together the status of a home plan for a
particular incarceration stint. The time comp team in AZ suggested we
look at only the highest `HOME_PLAN_DETAIL_ID` as that is likely the
most recent one. This PR filters the home plan preprocessed view to only
store information for rows where the `HOME_PLAN_DETAIL_ID` has the
highest numerical value for that particular incarceration stint.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 44a1e4e6e2456549cc220192cb70b6dc854eb59a